refactor(Game): change folders
