+++ src/app/app-routing.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import {AppComponent} from './app.component';
-import { GameComponent } from './game/game.component';
+import { GameComponent } from './components/game/game.component';
 import { HomeComponent } from './pages/home/home.component';
 import { LoginComponent } from './pages/login/login.component';
 import { MyGamesComponent } from './pages/my-games/my-games.component';
