+++ src/app/app.module.ts
@@ -18,7 +18,7 @@ import { NgxLoadingModule } from 'ngx-loading';
 import { AuthInterceptor } from './interceptors/auth.interceptor';
 import { RegisterGameComponent } from './pages/register-game/register-game.component';
 import { AuthGuard } from './services/auth.guard';
-import { GameComponent } from './game/game.component';
+import { GameComponent } from './components/game/game.component';
 import { MatButtonModule } from '@angular/material/button';
 import { NgxStarRatingModule } from 'ngx-star-rating';
 import { NgxInputStarRatingModule } from 'ngx-input-star-rating';
