+++ src/app/components/game/game.component.scss
