+++ src/app/components/game/game.component.ts
@@ -3,8 +3,8 @@ import { FormBuilder, FormGroup } from '@angular/forms';
 import { ActivatedRoute } from '@angular/router';
 import { tap } from 'rxjs/operators';
 import Swal from 'sweetalert2';
-import { Evaluation, EvaluationService } from '../services/evaluation.service';
-import { GameService } from '../services/game.service';
+import { Evaluation, EvaluationService } from '../../services/evaluation.service';
+import { GameService } from '../../services/game.service';
 
 @Component({
   selector: 'app-game-component',
